# gateway/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Описание эндпоинтов сервиса анонимизации данных",
        "title": "Anonix Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/isAdmin": {
            "post": {
                "description": "По переданному user_id возвращает, является ли пользователь администратором",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Проверить админство пользователя",
                "parameters": [
                    {
                        "description": "ID пользователя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.IsAdminSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.IsAdminRespSchema"
                        }
                    },
                    "400": {
                        "description": "invalid request body"
                    },
                    "500": {
                        "description": "failed to check if user is admin"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "description": "Возвращает роли авторизованного пользователя из токена доступа",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Получить роли текущего пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetUserRolesRespSchema"
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Принимает refresh token, возвращает новый access и refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Обновить токены",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.RefreshSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.RefreshRespSchema"
                        }
                    },
                    "400": {
                        "description": "invalid request body / failed to refresh token"
                    },
                    "500": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/auth/signIn": {
            "post": {
                "description": "Возвращает токен доступа и токен обновления",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация",
                "parameters": [
                    {
                        "description": "Данные для авторизации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.LoginRespSchema"
                        }
                    },
                    "400": {
                        "description": "invalid request body"
                    },
                    "401": {
                        "description": "invalid credentials"
                    },
                    "500": {
                        "description": "failed to log in"
                    }
                }
            }
        },
        "/auth/signUp": {
            "post": {
                "description": "Возвращает ID пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Зарегистрировать пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.RegisterSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.RegisterRespSchema"
                        }
                    },
                    "400": {
                        "description": "invalid request body"
                    },
                    "409": {
                        "description": "user with same login already exists"
                    },
                    "500": {
                        "description": "failed to register user"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.GetUserRolesRespSchema": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.RoleSchema"
                    }
                }
            }
        },
        "schemas.IsAdminRespSchema": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "boolean"
                }
            }
        },
        "schemas.IsAdminSchema": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "schemas.LoginRespSchema": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.RoleSchema"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "schemas.LoginSchema": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "schemas.RefreshRespSchema": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "schemas.RefreshSchema": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "schemas.RegisterRespSchema": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "schemas.RegisterSchema": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
        "schemas.RoleSchema": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}